+++ services/agent/tests/unit/conftest.py
@@ -1,16 +1,27 @@
 # pylint: disable=redefined-outer-name
 # pylint: disable=unused-argument
 
-from collections.abc import AsyncIterator
+import asyncio
+from collections.abc import AsyncIterable, AsyncIterator, Awaitable, Callable
+from contextlib import suppress
+from pathlib import Path
+from uuid import uuid4
 
+import aiodocker
 import pytest
+from aiodocker.containers import DockerContainer
+from aiodocker.volumes import DockerVolume
 from asgi_lifespan import LifespanManager
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
+from models_library.projects import ProjectID
+from models_library.projects_nodes_io import NodeID
 from models_library.services_types import RunID
+from models_library.users import UserID
 from pytest_simcore.helpers.monkeypatch_envs import EnvVarsDict, setenvs_from_dict
 from settings_library.rabbit import RabbitSettings
 from simcore_service_agent.core.application import create_app
+from utils import VOLUMES_TO_CREATE, get_source
 
 
 @pytest.fixture
@@ -48,3 +59,114 @@ def test_client(initialized_app: FastAPI) -> TestClient:
 @pytest.fixture
 def run_id() -> RunID:
     return RunID.create()
+
+
+@pytest.fixture
+def project_id() -> ProjectID:
+    return uuid4()
+
+
+@pytest.fixture
+def user_id() -> UserID:
+    return 1
+
+
+@pytest.fixture
+def volumes_path(tmp_path: Path) -> Path:
+    return tmp_path / "volumes"
+
+
+@pytest.fixture
+async def create_dynamic_sidecar_volume(
+    run_id: RunID,
+    project_id: ProjectID,
+    swarm_stack_name: str,
+    user_id: UserID,
+    volumes_path: Path,
+) -> AsyncIterable[Callable[[NodeID, bool, str, Path | None], Awaitable[str]]]:
+    volumes_to_cleanup: list[DockerVolume] = []
+    containers_to_cleanup: list[DockerContainer] = []
+
+    async with aiodocker.Docker() as docker_client:
+
+        async def _(
+            node_id: NodeID,
+            in_use: bool,
+            volume_name: str,
+            path_to_copy: Path | None = None,
+        ) -> str:
+            source = get_source(run_id, node_id, volumes_path / volume_name)
+            volume = await docker_client.volumes.create(
+                {
+                    "Name": source,
+                    "Labels": {
+                        "node_uuid": f"{node_id}",
+                        "run_id": run_id,
+                        "source": source,
+                        "study_id": f"{project_id}",
+                        "swarm_stack_name": swarm_stack_name,
+                        "user_id": f"{user_id}",
+                    },
+                }
+            )
+            volumes_to_cleanup.append(volume)
+
+            if in_use:
+                container = await docker_client.containers.run(
+                    config={
+                        "Cmd": ["/bin/ash", "-c", "sleep 10000"],
+                        "Image": "alpine:latest",
+                        "HostConfig": {"Binds": [f"{volume.name}:{volumes_path}"]},
+                    },
+                    name=f"using_volume_{volume.name}",
+                )
+                await container.start()
+                containers_to_cleanup.append(container)
+
+                # copy files form fodler if it exists
+                if path_to_copy:
+                    await asyncio.gather(
+                        *(
+                            container.exec(
+                                [
+                                    "sh",
+                                    "-c",
+                                    f"echo '{file_path.read_text()}' > "
+                                    "f/{volumes_path}/{file_path.name}",
+                                ]
+                            )
+                            for file_path in path_to_copy.rglob("*")
+                        )
+                    )
+
+            return source
+
+        yield _
+
+        for container in containers_to_cleanup:
+            with suppress(aiodocker.DockerError):
+                await container.delete(force=True)
+        for volume in volumes_to_cleanup:
+            with suppress(aiodocker.DockerError):
+                await volume.delete()
+
+
+@pytest.fixture
+def create_dynamic_sidecar_volumes(
+    create_dynamic_sidecar_volume: Callable[
+        [NodeID, bool, str, Path | None], Awaitable[str]
+    ]
+) -> Callable[[NodeID, bool], Awaitable[set[str]]]:
+    async def _(
+        node_id: NodeID, in_use: bool, path_to_copy: Path | None = None
+    ) -> set[str]:
+        volume_names: set[str] = set()
+        for volume_name in VOLUMES_TO_CREATE:
+            name = await create_dynamic_sidecar_volume(
+                node_id, in_use, volume_name, path_to_copy
+            )
+            volume_names.add(name)
+
+        return volume_names
+
+    return _
